+++ gridstatus/base.py
@@ -30,6 +30,35 @@ class Markets(Enum):
         return True
 
 
+class InterconnectionQueueStatus(Enum):
+    """Interconnection queue types"""
+
+    ACTIVE = "Active"
+    WITHDRAWN = "Withdrawn"
+    COMPLETED = "Completed"
+
+
+_interconnection_columns = [
+    "Queue ID",
+    "Project Name",
+    "Interconnecting Entity",
+    "County",
+    "State",
+    "Interconnection Location",
+    "Transmission Owner",
+    "Generation Type",
+    "Capacity (MW)",
+    "Summer Capacity (MW)",
+    "Winter Capacity (MW)",
+    "Queue Date",
+    "Status",
+    "Proposed Completion Date",
+    "Withdrawn Date",
+    "Withdrawal Comment",
+    "Actual Completion Date",
+]
+
+
 class ISOBase:
     markets = []
     status_homepage = None
@@ -76,6 +105,9 @@ class ISOBase:
         raise NotImplemented()
 
     def get_storage(self, date, end=None, verbose=False):
+        raise NotImplementedError()
+
+    def get_interconnection_queue(self):
         raise NotImplementedError()
 
     def _latest_lmp_from_today(self, market, locations, **kwargs):